mha fix

+++ mha_benchmark.py
@@ -9,7 +9,7 @@ def MHA_CPU(embed_dim, batch_size, seq_len, num_heads):
     return time.time() - start
 
 def MHA_GPU(embed_dim, batch_size, seq_len, num_heads):
-    mha = torch.nn.MultiheadAttention(embed_dim=embed_dim, num_heads=num_heads, batch_first=True).cuda()
+    mha = torch.nn.MultiheadAttention(embed_dim=embed_dim, num_heads=num_heads, batch_first=True).cuda().half()
     x = torch.rand(batch_size, seq_len, embed_dim).cuda().half()
 
     for _ in range(3):
